Add a new validation and a slider for the mine density to configuration

+++ client/src/main/java/tk/aakado/multisweeper/client/views/configuration/ConfigurationViewModel.java
@@ -2,9 +2,11 @@ package tk.aakado.multisweeper.client.views.configuration;
 
 import de.saxsys.mvvmfx.ViewModel;
 import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.ListProperty;
 import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleListProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -14,7 +16,7 @@ import javafx.collections.FXCollections;
 public class ConfigurationViewModel implements ViewModel, ConfigurationNotificator {
 
     private BooleanProperty admin = new SimpleBooleanProperty(false);
-    private IntegerProperty mineDensity = new SimpleIntegerProperty();
+    private DoubleProperty mineDensity = new SimpleDoubleProperty();
     private IntegerProperty fieldWidth = new SimpleIntegerProperty();
     private IntegerProperty fieldHeight = new SimpleIntegerProperty();
     private StringProperty password = new SimpleStringProperty();
@@ -48,7 +50,7 @@ public class ConfigurationViewModel implements ViewModel, ConfigurationNotificat
     }
 
 
-    public IntegerProperty mineDensityProperty() {
+    public DoubleProperty mineDensityProperty() {
         return mineDensity;
     }
